requirements.txt
flask
flask-cors
requests
dateparser
pytz
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
telethon 
python-dotenv